+++ source/js/mgr/widgets/resourcematrix.grid.js
@@ -64,9 +64,6 @@ Babel.grid.ResourceMatrix = function (config) {
     var cm = new Ext.ux.grid.LockingColumnModel({
         columns: columns
     });
-    var view = new Ext.ux.grid.LockingGridView({
-        syncHeights: true
-    });
 
     Ext.apply(config, {
         url: Babel.config.connectorUrl,
@@ -131,227 +128,13 @@ Babel.grid.ResourceMatrix = function (config) {
 };
 Ext.extend(Babel.grid.ResourceMatrix, MODx.grid.Grid, {
     linkTranslation: function (ctx, id) {
-        var window = MODx.load({
-            xtype: 'modx-window',
-            title: _('babel.link_translation'),
-            url: Babel.config.connectorUrl,
-            baseParams: {
-                action: 'mgr/resource/link',
-                context: ctx,
-                id: id
-            },
-            listeners: {
-                success: {
-                    fn: function (r) {
-                        MODx.msg.status({
-                            title: _('success'),
-                            message: r.message || _('save_successful')
-                        });
-                        this.hideMask();
-                        this.refresh();
-                    },
-                    scope: this
-                },
-                failure: {
-                    fn: this.hideMask,
-                    scope: this
-                },
-                beforeSubmit: {
-                    fn: this.loadMask,
-                    scope: this
-                }
-            },
-            fields: [{
-                xtype: 'textfield',
-                fieldLabel: _('context'),
-                anchor: '100%',
-                disabled: true,
-                emptyText: ctx
-            }, {
-                xtype: 'modx-field-parent-change',
-                fieldLabel: _('babel.select_tree_node'),
-                id: '',
-                name: 'target-combo',
-                anchor: '100%',
-                end: function (parent) {
-                    var tree = Ext.getCmp('modx-resource-tree');
-                    if (!tree) {
-                        return;
-                    }
-                    parent.display = parent.display || parent.value;
-                    if (parent.context !== ctx) {
-                        return;
-                    }
-                    tree.removeListener('click', this.handleChangeParent, this);
-                    tree.on('click', tree._handleClick, tree);
-                    tree.disableHref = false;
-                    window.fp.getForm().findField('target').setValue(parent.value);
-                    window.fp.getForm().findField('page_id').setValue(null);
-                    this.setValue(parent.display);
-                    this.oldValue = false;
-                },
-                handleChangeParent: function (node, e) {
-                    var tree = Ext.getCmp('modx-resource-tree');
-                    if (!tree) {
-                        return false;
-                    }
-                    tree.disableHref = true;
-                    var id = node.id.split('_');
-                    id = id[1];
-                    if (id === this.config.currentid) {
-                        MODx.msg.alert('', _('resource_err_own_parent'));
-                        return false;
-                    }
-                    this.fireEvent('end', {
-                        value: node.attributes.type !== 'modContext' ? id : node.attributes.pk,
-                        display: Ext.util.Format.stripTags(node.text),
-                        context: node.attributes.ctx
-                    });
-                    e.preventDefault();
-                    e.stopEvent();
-                    return true;
-                }
-            }, {
-                xtype: 'modx-combo',
-                fieldLabel: _('babel.target_pagetitle'),
-                name: 'page_id',
-                anchor: '100%',
-                url: Babel.config.connectorUrl,
-                baseParams: {
-                    action: 'mgr/resource/getlist',
-                    context: ctx,
-                    combo: true
-                },
-                displayField: 'pagetitle',
-                valueField: 'id',
-                fields: ['id', 'pagetitle'],
-                editable: true,
-                typeAhead: true,
-                forceSelection: true,
-                listeners: {
-                    select: {
-                        fn: function (combo, record) {
-                            var val = combo.getValue();
-                            if (val === '' || val === 0 || val === '&nbsp;') {
-                                combo.setValue(null);
-                            } else {
-                                window.fp.getForm().findField('target').setValue(record.get('id'));
-                            }
-                            window.fp.getForm().findField('target-combo').reset();
-                        },
-                        scope: this
-                    },
-                    blur: {
-                        fn: function (combo) {
-                            var val = combo.getValue();
-                            if (val === '' || val === 0 || val === '&nbsp;') {
-                                combo.setValue(null);
-                            }
-                        },
-                        scope: this
-                    }
-                }
-            }, {
-                fieldLabel: _('babel.target_id'),
-                xtype: 'numberfield',
-                name: 'target',
-                anchor: '100%',
-                enableKeyEvents: true,
-                listeners: {
-                    keyup: {
-                        fn: function () {
-                            window.fp.getForm().findField('target-combo').reset();
-                            window.fp.getForm().findField('page_id').reset();
-                        },
-                        scope: this
-                    }
-                }
-            }, {
-                xtype: 'xcheckbox',
-                boxLabel: _('babel.copy_tv_values'),
-                name: 'copy-tv-values'
-            }, {
-                xtype: 'xcheckbox',
-                boxLabel: _('babel.sync_linked_tranlations'),
-                name: 'sync-linked-tranlations',
-                checked: true
-            }]
-        });
-        window.reset();
-        window.show();
+        Babel.linkTranslation(ctx, id, this);
     },
     unlinkTranslation: function (ctx, id, target) {
-        this.loadMask();
-        ctx = ctx || '';
-        target = parseInt(target) || 0;
-        var text = (target === 0) ?
-            _('babel.unlink_all_translations_confirm') :
-            _('babel.unlink_translation_confirm', {
-                context: ctx,
-                id: id
-            });
-        return MODx.msg.confirm({
-            title: _('confirm'),
-            text: text,
-            url: Babel.config.connectorUrl,
-            params: {
-                action: 'mgr/resource/unlink',
-                id: id,
-                context: ctx,
-                target: target
-            },
-            listeners: {
-                success: {
-                    fn: function (r) {
-                        MODx.msg.status({
-                            title: _('success'),
-                            message: r.message || _('save_successful')
-                        });
-                        this.hideMask();
-                        this.refresh();
-                    },
-                    scope: this
-                },
-                failure: {
-                    fn: this.hideMask,
-                    scope: this
-                },
-                cancel: {
-                    fn: this.hideMask,
-                    scope: this
-                }
-            }
-        });
+        return Babel.unlinkTranslation(ctx, id, target, this);
     },
     createTranslation: function (ctx, id) {
-        this.loadMask();
-        return MODx.msg.confirm({
-            title: _('confirm'),
-            text: _('babel.create_translation_confirm', {context: ctx, id: id}),
-            url: Babel.config.connectorUrl,
-            params: {
-                action: 'mgr/resource/duplicate',
-                context_key: ctx,
-                id: id
-            },
-            listeners: {
-                success: {
-                    fn: function (r) {
-                        this.hideMask();
-                        MODx.loadPage('resource/update', 'id=' + r.object.id);
-                    },
-                    scope: this
-                },
-                failure: {
-                    fn: this.hideMask,
-                    scope: this
-                },
-                cancel: {
-                    fn: this.hideMask,
-                    scope: this
-                }
-            }
-        });
+        Babel.createTranslation(ctx, id)
     },
     clearFilter: function () {
         var store = this.getStore();
@@ -373,26 +156,16 @@ Ext.extend(Babel.grid.ResourceMatrix, MODx.grid.Grid, {
         this.refresh();
         return true;
     },
-    loadMask: function () {
-        if (!this.overlayMask) {
-            var domHandler = Ext.getBody().dom;
-            this.overlayMask = new Ext.LoadMask(domHandler, {
-                msg: _('babel.please_wait')
-            });
-        }
-        this.overlayMask.show();
-    },
-    hideMask: function () {
-        if (this.overlayMask) {
-            this.overlayMask.hide();
-        }
-    },
     buttonColumnAllRenderer: function () {
         return this.buttonColumnTpl.apply({
             action_buttons: [{
+                className: 'create-multiple',
+                icon: 'plus-square',
+                text: _('babel.create_multiple_translations')
+            }, {
                 className: 'unlink-all',
                 icon: 'chain-broken',
-                text: _('babel.unlink')
+                text: _('babel.unlink_all')
             }]
         });
     },
@@ -446,6 +219,9 @@ Ext.extend(Babel.grid.ResourceMatrix, MODx.grid.Grid, {
                 case 'unlink-all':
                     this.unlinkTranslation('', record.get('id'), 0)
                     break;
+                case 'create-multiple':
+                    this.createTranslation('', record.get('id'));
+                    break;
                 case 'create':
                     this.createTranslation(t.dataset.ctx, record.get('id'));
                     break;
@@ -454,7 +230,6 @@ Ext.extend(Babel.grid.ResourceMatrix, MODx.grid.Grid, {
                     break;
                 case 'update':
                     MODx.loadPage('resource/update', 'id=' + t.dataset.target);
-                    console.log(record);
                     break;
                 case 'unlink':
                     this.unlinkTranslation(t.dataset.ctx, record.get('id'), t.dataset.target);